redesign tests so that peak mem test can be run by itself
