+++ .cargo/config
@@ -5,7 +5,7 @@
 # It can only be run from 10x Genomics because it uses a bunch of data files that we don't
 # distribute.
 
-t = "test --bin traceback1 --test enclone_test1 --test enclone_test2 -- --nocapture"
+t = "test --bin traceback1 --test enclone_test1 --test enclone_test2 --test enclone_test_peak_mem -- --nocapture"
 
 # Note that "./test" exercises this plus a few more tests.
