+++ enclone_main/tests/enclone_test2.rs
@@ -655,117 +655,7 @@ fn test_subset_json() {
 
 // ▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓
 
-// 25. Test peak memory.  This is designed for one server, bespin1.
-//
-// You can run this test along by typing
-// cargo test --manifest-path enclone_main/Cargo.toml --features mem -- --nocapture
-// from the root of the repo.
-// However this also runs the doc tests.  There must be a better way.
-
-#[cfg(not(feature = "basic"))]
-#[cfg(not(feature = "cpu"))]
-#[cfg(feature = "mem")]
-#[test]
-fn test_peak_memory() {
-    PrettyTrace::new().on();
-
-    // Specify mem requirements.
-
-    let dataset = "BCR=123085";
-    let expected_mb = 364.8;
-
-    let max_percent_dev = 0.5;
-
-    // Only run internally.
-
-    let mut internal_run = false;
-    for (key, value) in env::vars() {
-        if key.contains("TELEPORT") && value.contains("10xgenomics.com") {
-            internal_run = true;
-        }
-    }
-    if !internal_run {
-        return;
-    }
-
-    // Define function to get mem.
-
-    fn mem_val(dataset: &str) -> f64 {
-        let new = Command::new(env!("CARGO_BIN_EXE_enclone"))
-            .arg(&dataset)
-            .args(&["NOPRINT", "COMP"])
-            .output()
-            .expect(&format!("failed to execute peak mem test"));
-        if new.status.code() != Some(0) {
-            eprint!(
-                "\nenclone_site_examples: nonzero status code for peak mem test, stderr =\n{}",
-                strme(&new.stderr),
-            );
-            std::process::exit(1);
-        }
-        let out = strme(&new.stdout);
-        let mut mem = None;
-        for line in out.lines() {
-            if line.starts_with("peak mem usage = ") && line.ends_with(" MB") {
-                mem = Some(line.between("peak mem usage = ", " MB").force_f64());
-            }
-        }
-        if mem.is_none() {
-            eprintln!("\nfailed to find mem usage for peak mem\n");
-            std::process::exit(1);
-        }
-        mem.unwrap()
-    }
-
-    // Run the test.
-
-    let mut ok = false;
-    let mut mems = Vec::<f64>::new();
-    for _tries in 0..3 {
-        let mem = mem_val(&dataset);
-        mems.push(mem);
-        let dev = 100.0 * (mem - expected_mb).abs() / expected_mb;
-        if dev <= max_percent_dev {
-            ok = true;
-            break;
-        }
-    }
-    if !ok {
-        for _tries in 3..10 {
-            mems.push(mem_val(&dataset));
-        }
-    }
-    let mut mean = 0.0;
-    for i in 0..mems.len() {
-        mean += mems[i];
-    }
-    mean /= mems.len() as f64;
-    let dev = 100.0 * (mean - expected_mb).abs() / expected_mb;
-    let msg = format!(
-        "Peak mem of {:.1} MB observed, which differs from expected value of {} by {:.2}%.",
-        mean, expected_mb, dev
-    );
-    if !ok && dev > max_percent_dev {
-        eprintln!("\n{}\n", msg);
-        eprintln!("observed values:\n");
-        for i in 0..mems.len() {
-            eprintln!("{:.1}", mems[i]);
-        }
-        eprintln!(
-            "Please note that this test was designed to work correctly from a single server\n\
-            named bespin1.  If you're running from a different server, the expected memory value\n\
-            may need to be changed.  This might also be the case if that server was changed.\n\
-            Otherwise, your options are:\n\
-            1. Change the value of expected_mb in enclone_test.rs to {:.1}.\n\
-            2. Optimize to reduce mem usage (if value exceeds expected).\n\
-            3. Repeat the test, but this is very unlikely to succeed.\n",
-            mean
-        );
-        std::process::exit(1);
-    } else {
-        println!("{}", msg);
-    }
-}
+// 25. MOVED TO SEPARATE FILE
 
 // ▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓
